Move full focus toggle to notebook header
